+++ appseed/ca/os/os_cross_windows_crt.cpp
@@ -19,7 +19,7 @@ CLASS_DECL_ca errno_t _ctime64_s(char * buf, int32_t iSize, const time_t * timer
 
    int32_t iLen;
 
-   if(((iLen = strlen(ret)) + 1) > iSize)
+   if(((iLen = (int32_t) strlen(ret)) + 1) > iSize)
       return ERANGE;
 
    strncpy(buf, ret, min(iSize, iLen));
